Moving rendering files around for better structure
