+++ example/demo_gen/demo/rendering/FixedDecimalFormatter_formatWrite.html
@@ -3,10 +3,10 @@
     <head>
         <meta charset="utf-8"/>
         <title>FixedDecimalFormatter.formatWrite</title>
-        <script src="../runtime.mjs" type="module"></script>
+        <script src="runtime.mjs" type="module"></script>
     </head>
     <body>
-        <iframe src="../template.html" id="templates" hidden></iframe>
+        <iframe src="template.html" id="templates" hidden></iframe>
         <terminus-render func="FixedDecimalFormatter.formatWrite"></terminus-render>
     </body>
 </html>
