+++ tool/templates/demo_gen/default_renderer/termini.html.jinja
@@ -3,10 +3,10 @@
     <head>
         <meta charset="utf-8"/>
         <title>{{t.type_name}}.{{t.function_name}}</title>
-        <script src="../runtime.mjs" type="module"></script>
+        <script src="runtime.mjs" type="module"></script>
     </head>
     <body>
-        <iframe src="../template.html" id="templates" hidden></iframe>
+        <iframe src="template.html" id="templates" hidden></iframe>
         <terminus-render func="{{t.type_name}}.{{t.function_name}}"></terminus-render>
     </body>
 </html>
